+++ src/clj/py99/routes/services.clj
@@ -7,26 +7,6 @@
    [py99.utils :as u]
    [ring.util.http-response :as response]))
 
-;;;;;;;;;;;;;;;;;;;;;;
-;; from home.clj
-;; (defn- before? [s1 s2]
-;;   ;; 2022-10-20 s/</<=/
-;;   (<= (compare s1 s2) 0))
-
-;; (defn- count-up [m]
-;;   (reduce + (map :count m)))
-
-;; (defn bin-count
-;;   [data bin]
-;;   (loop [data data bin bin ret []]
-;;     (if (empty? bin)
-;;       ret
-;;       (let [g (group-by #(before? (:create_at %) (first bin)) data)
-;;             f (g true)
-;;             s (g false)]
-;;         (recur s (rest bin) (conj ret (count-up f)))))))
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-
 (defn fetch-problem
   [{{:keys [n]} :path-params}]
   ;; (prn "fetch-problem" n)
@@ -53,9 +33,9 @@
   (let [zeros (count (filter #(= 0 %) col))]
     (* (apply + col) (- 6 zeros))))
 
-(defn s-point-login-date
+(defn s-point
   "calc `login`s s-point from star to until `date`."
-  [{{:keys [login date]} :path-params}]
+  [login date]
   (let [date-count (db/answers-by-login-date
                     {:login login :date date})
         dc (apply merge (for [mm date-count]
@@ -68,12 +48,14 @@
                 (map s)
                 (apply +))]
     (log/debug "s-point-login-date" py99 sp)
-    ;; Content-Type: application/json; charset=utf-8 で返る。
-    ;; middleware が賢いのか？
     (response/ok {:login login
                   :date date
                   :py99 py99
                   :s sp})))
+
+(defn s-point-login-date
+  [{{:keys [login date]} :path-params}]
+  (s-point login date))
 
 (defn points [{{:keys [login]} :path-params}]
   (response/ok
@@ -102,30 +84,30 @@
   (response/ok
    (assoc (db/solved-by {:login login}) :login login)))
 
-(defn py99!
-  [{{:keys [secret login col pt]} :params}]
-  (if (= secret (System/getenv "PY99_PASSWORD"))
-    (let [pt (Integer/parseInt pt)]
-      (case col
-        "e1" (db/update-e1! {:login login :pt pt})
-        "e2" (db/update-e2! {:login login :pt pt})
-        "e3" (db/update-e3! {:login login :pt pt})
-        "e4" (db/update-e4! {:login login :pt pt})
-        "e5" (db/update-e5! {:login login :pt pt})
-        "wil" (db/update-wil! {:login login :pt pt})
-        "py99" (db/update-py99! {:login login :pt pt})
-        "comm" (db/update-comm! {:login login :pt pt})
-        "goal-in" (db/update-goal! {:login login :pt pt})
-        "seven-four" (db/update-seven-four! {:login login :pt pt})
-        "default")
-      (response/ok {:login login
-                    :col col
-                    :pt pt
-                    :secret secret}))
-    (response/bad-request {:login login
-                           :col col
-                           :pt pt
-                           :secret secret})))
+; (defn py99!
+;   [{{:keys [secret login col pt]} :params}]
+;   (if (= secret (System/getenv "PY99_PASSWORD"))
+;     (let [pt (Integer/parseInt pt)]
+;       (case col
+;         "e1" (db/update-e1! {:login login :pt pt})
+;         "e2" (db/update-e2! {:login login :pt pt})
+;         "e3" (db/update-e3! {:login login :pt pt})
+;         "e4" (db/update-e4! {:login login :pt pt})
+;         "e5" (db/update-e5! {:login login :pt pt})
+;         "wil" (db/update-wil! {:login login :pt pt})
+;         "py99" (db/update-py99! {:login login :pt pt})
+;         "comm" (db/update-comm! {:login login :pt pt})
+;         "goal-in" (db/update-goal! {:login login :pt pt})
+;         "seven-four" (db/update-seven-four! {:login login :pt pt})
+;         "default")
+;       (response/ok {:login login
+;                     :col col
+;                     :pt pt
+;                     :secret secret}))
+;     (response/bad-request {:login login
+;                            :col col
+;                            :pt pt
+;                            :secret secret})))
 
 (defn recents [{{:keys [n]} :path-params}]
   (println "recents" n)
@@ -136,12 +118,13 @@
   []
   ["/api" {:middleware [middleware/wrap-formats]}
    ["/actions/:login/:date" {:get actions?}]
-   ["/comm/:login" {:get comm}]
-   ["/goal-in/:login" {:get goal-in}]
-   ["/points/:login" {:get points}]
-   ["/problem/:n" {:get fetch-problem}]
-   ["/pt/:login" {:get pt}]
-   ["/py99/:login" {:get py99}]
+   ; ["/comm/:login" {:get comm}]
+   ; ["/goal-in/:login" {:get goal-in}]
+   ; ["/points/:login" {:get points}]
+   ; ["/problem/:n" {:get fetch-problem}]
+   ; ["/pt/:login" {:get pt}]
+   ; ["/py99/:login" {:get py99}]
    ["/s/:login/:date" {:get s-point-login-date}]
-   ;; ["/recents/:n" {:get recents}]
-   ["/py99" {:post py99!}]])
+   ; ["/recents/:n" {:get recents}]
+   ; ["/py99" {:post py99!}]
+   ])