+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 
 ## v1.17-SNAPSHOT / 2024-11-28
 
+- endpoint /s-point
 - added: /api/recents/:n
 
 ## v1.16.1142 / 2024-11-28